GeneratorBackend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
transformers==4.52.4
torch==2.7.1
openai==1.90.0
httpx==0.28.1
python-dotenv==1.1.1